Move GitHubResponse to top-level class
* Simplifies GitHubApi by de-cluttering it
* No longer need to import inner class
